moved the NFFS_AREA_MAX to the bsp since it must match between a bootloader and another project sharing flash
made the hal flash for arduino zero use 1k sector size even though the pages are 256 byte.

fixed boot.c to ensure that we have enough area descriptors for the boot and added assert sanity to check in new builds

passed the offset of the area descriptors to nffs_detect so that it can find the real flash filesystem from the list
of area descriptors.

switch slinky to use baselibc

as of this checkin slinky works for arduino zero for image loading and booting