This closes #11. Support building the 'sim' with optimization levels higher than -O0. The higher optimization levels emit useful warnings that are masked at -O0.
The basic problem with the implementation of 'os_arch_task_stack_init()'
is that we are switching stacks from underneath the compiler without any
ability to influence how the compiler uses the stack. Thus when setjmp()
returns the compiler doesn't know that the stack frame it allocated for
local variables doesn't exist anymore on the new stack.

This happens to work with the -O0 optimization level because the compiler
addresses local variables using the frame pointer (%ebp). Higher optimization
levels use offsets from the stack pointer and immediately expose the issue
mentioned above.

Fix this by switching stacks in 'os_arch_frame_init()' which is implemented
in assembly so we can control the use of stack after setjmp() returns.