Restructure signal handling logic in 'sim' to make it similar to interrupts on real hardware. This allows other interrupt sources beyond just timers in sim (e.g. reads/write using aio).
Switch the interval timer from ITIMER_VIRTUAL to ITIMER_REAL. This allows us
to yield when we are idle and still get track passage of time faithfully (to
do this with ITIMER_VIRTUAL requires busy waiting in the idle loop). This
change coupled with using 'sigsuspend()' in the idle loop gets the CPU
utilization from 100% to 4% at a timer frequency of 1000hz.

A detailed explanation of how context switching works with signals follows:

A voluntary context switch ends up with 'os_sched()' entering the critical
section and doing a 'longjmp()' to the task being scheduled. The new task
returns via 'os_sched()' and exits the critical section on its way out.

An involuntary context switch starts with a signal handler that enters a
critical section implicitly because all signals are masked when the handler
is executing. This in turn means that 'os_sched()' is running in a nested
critical section. The interrupted task remains on the run queue and resumes
execution in the signal handler when it runs again. The cpu executes with
signals/interrupts blocked until the handler returns at which time the
interrupted task gets control again.