This is a second version of the hal_pwm API.
Its based on a discussion from the devlist.

It contains two different ways to "set" the PWM waveform, a
simple way with duty cycle and a more advanced way with clock ticks.

It also adds methods to query the resolution(bits) and clock
frequency of the underlying device when use the more advanced
waveform control.