Enter tickless regime during idle on NRF52.
'os_arch_idle(ticks)' now takes an argument to indicate how long before
the OS has anything useful to schedule. The architecture specific function
is allowed to suspend execution for up to those many ticks. Note that
this function may return earlier if another interrupt wakes up the CPU.
Note that 'ticks' may be zero in which case the CPU stays in the periodic
timer regime.

The idle stack size needed a bump to account for the increase in stack
depth.