bletiny - fix "set addr" command.
Background: The host is transitioning over to using its own concept of
its device address, rather than relying on the controller.  This
transition is necessary to support host-only setups.  As part of this
transition, the security manager code uses the host's conception of its
address, and does not use the controller's address at all.

This commit: Prior to this change, the "set addr" command only set the
controller's global address field (g_dev_addr).  Now, this command sets
both the controller's and host's address field.

This command is problematic for a number of reasons:
    * Changing one's address mid-connection is not compliant with the
      BLE spec.
    * The command assumes the nimble controller is built in to the app
      (combined host-controller).
    * Changing the host address requires the inclusion of a private
      net/nimble/host header file, which should not be done.

As a consequence, this command should be removed soon.  Unfortunately,
there is not currently a convenient way to set one's address, aside from
rebuilding bletiny with a different hardcoded value.