Changing has_objects to wr_commas and fixing it
For json a comma should be present at the end of every object if there
are multiple objects. Also, there shouldn't be trailing commas.

Problem: The je_has_objects flag would track commas as there wasn't a specific
way to do that with the json lib. It wasn't adding commas at the right
places.

Solution: A comma gets added at the right places now. Particularly, at
the end of each object and the flag gets reset each time a comma gets
added.

This should fix the problem we see with nested json encoding.
Also changing the name to be more specific.