Make task context switches in sim behave similarly to real hardware.
Prior to the change context switching in sim happened "inline" as opposed
to on hardware where it would trigger a special interrupt and the context
switching happened in context of the interrupt handler.

This causes subtle differences in how tasks are scheduled inside sim.
For e.g. 'os_callout_tick()' is called from the timer handler with
interrupts disabled. If a callout fires this function will post to an
eventq and if a task is sleeping on it then it will be made runnable.

On 'sim' the context switch happened immediately (i.e. from the timer
handler context to the task that became runnable) whereas on real hardware
all it would do is trigger the special interrupt. Note that the interrupt
will not be recognized until the timer handler re-enables interrupts.

The context switching behavior is now done similarly on sim by posting
a signal to ourselves. The signal used is SIGURG because gdb treats this
signal specially and forwards it to the process without any additional
configuration (as opposed to gdb's treatment of SIGUSR1 which stops the
process).