ble host - major changes.
These changes are detailed in this thread on the mynewt dev list:
http://mail-archives.apache.org/mod_mbox/incubator-mynewt-dev/201604.mbox/%3C20160417225734.GB8991%40iori.nightmare-heaven.no-ip.biz%3E

Here is a summary of the changes:

1. All HCI command/acknowledgement exchanges are blocking.

    * FSM machinery goes away; controller response is indicated in the
      return code of the HCI send function.
    * Nearly all HCI failures are indicated to the application
      immediately, so there is no need for lots of mutexes and temporary
      copies of data structures.
    * API is simplified; operation results are indicated via a simple
      function return code.

2. The Nimble host is "taskless"

the Nimble host is now a "flat" library that runs in an application
task.  When the application initializes the host, it indicates which OS
event queue should be used for host-related events.  Host operations are
captured in OS_EVENT_TIMER events that the application task would need
to handle generically, as it most likely already does.