Fix for 64-bit types passed to *printf.
The va_arg macro from newlib seems to have issues with 64-bit values; it
was reading from the wrong memory location.  The temporary fix is to
pull the 64-bit number off the stack as a pair of 32-bit integers.

We need to look into why this is necessary.  It may just be an incorrect
setting specified when the compiler / newlib was built.