Don't do anything in signal handlers when process is blocked in 'sigsuspend()'.
When the OS is idling it may disable the periodic timer and enter the tickless
mode. The OS exits the tickless mode on arrival of an interrupt/signal (note
that OS time is incorrect when this happens). Thus it is important to call
'timer_handler()' to correct the OS time before executing anything else.