nffs - Fix for broken iteration through all objs.
The NFFS_HASH_FOREACH macro was broken because it did not account for
the fact that read operations can still modify the ordering in the hash.
In particular, lookups cause an object to be moved to the front of the
list to facilitate LRU.

Now, the macro stores a "next" pointer before executing the body of the
loop.