Fix bug in data length extension procedure
Changes were made to make sure that the hosts suggested data
sizes were never greater than the supported controller size.
That check was wrong (should have been less than or equal to) so
the controller would not perform the data length extension
procedure in some cases (when the maximum supported size was
requested).