MYNEWT-99: implement encryption pause procedure
This allows keys to be changed on an already encrypted connection.
The code was tested on a bmd300, nrf52dk and also a nrf51. The
nrf51 with encryption enabled was disconnecting and/or getting
MIC failures with large frames; was working fine for small
frames. We are investigating similar issues on the nrf51 so I
do not necessarily believe these to be encryption related which
is why I am committing the changes.