Fix scheduling bugs
There were two scheduling bugs that were uncovered during recent
testing. The first was that we should have removed the schedule
element prior to calling the connection start callback. The
second is that a currently scheduled connection could get moved
when a new connection is added. Both of these bugs could cause
connections to become disconnected if more than one connection
is ongoing.