json - Fix array-of-int decoding; add basic test.
Array decoding seems to be broken.  This is an inefficient and ugly fix
just to get it working.  We will need to rethink the semantics of the
function pointers that get used here (read_next, read_prev, and readn).
The fundamental problem is that the current offset is adjusted
automatically upon read when sometimes we just need to "peek" at the
current character.

The other issue is that we don't know how many characters an string
representation of an int occupies.  Currently we just read up to 64
characters and then backtrack until the end of the number.  It would be
better to read individual characters until we hit the end of the number.