BLE Host - Fix build when security disabled.
The host failed to compile when NIMBLE_OPT_SM was disabled.  This commit
fixes this build error.

The change which broke the build should never have been committed in the
first place.  That change is part of a larger set of changes that need
to be committed at the same time.  This was my screw up.  Sorry
everyone!