Modify on-flash structures in prep for lastblock change
Change on-flash data structures including the addition of ndi_lastblock_id
in the nffs_disk_inode struct.  This field will store the last block of each
file for fast restores. That change is not enabled in this set of changes,
it will come in a subsequent commit.
Other changes include:
* Deleting the per inode/block magic number (ndi_magic, ndb_magic)
* Modifying the code to read in disk objects during restore and in ffs2native.
* Incrementing the filesystem version number stored in struct nffs_disk-area
* Add "-s" flag to ffs2native to parse NFFS flash area images (V1 format for now)