Implement file lastblock caching in flash
This change stores the last written block for each file in the metadata
stored in flash. It facilitates restoring and mounting the filesystem
during initialization by removing the requirement to search every block in
flash to find the last block of each file.

This change requires the previous commit changing the on-flash data structures
which include a reference to the lastblock. This change is not backwards
compatible with previous on-flash version of NFFS, and there is a check
to make sure that the current version is used. Users upgrading to this version
of NFFS will need to reformat their flash devices - currently there is no
facility for in-place upgrades of existing filesystems.

One side-effect of this change is that inode updates are written to flash
each time a new block is added to the file. Other changes include additional
state flags with checking to provide additional resilience with minimal
space and compute overhead. There is also additional debugging tools in
nffs_test.c.