Legacy NFFS version support to standalone mode of ffs2native
Standalone mode (-s) takes an NFFS flash image file and prints the contents
without regards to a specific flash area layout. The layout is determined from
parsing the input file and this allows NFFS format image files from
any device to be printed, regardless of the HW flash layout.

This change adds support for V0 layouts - the version is automatically
detected from parsing the data structures in the image.