NimBLE - Host frees HCI cmd buf on ev alloc fail.
Old behavior: when ble_hci_transport_host_cmd_send() failed to allocate
an event for the HCI command, it tried to free the command buffer to the
g_hci_cmd_pool pool.  This became a problem when the host started using
a single statically-allocated buffer for all of its HCI commands.

New behavior: ble_hci_transport_host_cmd_send() doesn't free the buffer
on failure.  Instead, it is up to the caller to check the return code
and free the buffer on failure.