BLE Host - GATT procedures handled in wrong order.
This was a problem when the application initiated two concurrent GATT
procedures with the same op code.  In this case, when a response was
received, it was paired with the incorrect GATT procedure when the
callback was executed.

The fix is to append GATT procedures to the tail of the stailq rather
than the head.