att: use offset when preparing Read Blob Response
This remediates a bug wherein the offset of a Read Blob Request (from
the central) is ignored and the response always includes first MTU-1
bytes of the attribute value regardless of offset requested. As a
result, most hosts sit in an infinite loop incrementing the offset but
continually receiving the first MTU-1 bytes.

Note that this code has not been carefully tested for boundary
conditions.