Update ffs2native with bugfixes and CRC support
This fixes a bug in how V1 filesystems are parsed. Also, the CRC of each record
is verified and also that it's length does not extend beyond a flash area.
Records are printed with error messages as detected. This only occurs if there
is file corruption on the flash device.