Don't automatically truncate files in fsutil_write_file()
Prior to this change, fsutil_write_file() opened files for writing with the
truncate flag.  This caused all writes to delete the old version of the file,
create a new file and then update the file with the newly written data.
This caused 3 times the records written to the flash device (ie, the delete
record for the old file, the creation record for the new file, and the update
for the lastblock when the data was actually written).  Now only a new block is
generated and written to flash (with a higher sequence number if appropriate).
fsutil_write_file() is used extensively in the boot upgrade process.