Bug fixes and debug enhancements for nffs
These changes were made to address jira issue mynewt-343. A couple of bugs were
fixed include the removal of an incorrect assert in nffs_restore_inode(), and
resetting the sweep when a block is removed in nffs_restore_sweep(). Also,
nffs_stats() routine was added to track how many objects were parsed and are
in the hash table.

This is a partial fix for mynewt-343 - it passes all unit tests and has been
verified to improved the robustness during the boot image upgrade on
arduino-zero boards.