blecent - Fix memory leak on disconnect!
When a connection is terminated, the peer is removed from the peer list.
The peer object was getting freed, but not the discovered services,
characteristics, and descriptors.

Also, the discovery procedure was not getting aborted when the
application failed to add a GATT resource to a peer (e.g., due to memory
exhaustion).  This commit fixes this problem as well.