ble example apps: use dflt ble_hs cfg if possible
Prior to this change, the sampel apps set most of the ble_hs_cfg values
to somewhat arbitrary numbers.  Now, the apps leave most settings with
their default values.  The defaults are more appropriate in terms of the
apps' actual requirements.  This change also makes the apps easier to
copy and paste from for people getting started with NimBLE.