ATT: send a Read By Group Type Response in the case of BLE_HS_EMSGSIZE
Read By Group Type Response messages can raise BLE_HS_EMSGSIZE when the
response is large and the MTU is small. In this case the message built
by ble_att_svr_build_read_group_type_rsp() is valid and complete
(despite the error-looking return value), and should be sent to the
remote party.