BLE Host - Remove double indirection from tx API.
Prior to this change, data buffers for tx operations were passed as type
(struct os_mbuf **).  The second pointer was used so that the host could
set the caller's pointer to NULL, indicating that the mbuf had been
consumed and preventing a double-free.

This was useful, but confusing.  Using a single pointer is simpler and
consistent with the other networking libraries in core.  It is up to the
caller to read the function spec and understand that ownership of the
mbuf is transferred.