BLE Host - validate write upon rx of prep write
Before, upon receiving a prepare write request, the code was only
verifying that the specified attribute existed.  There are a number of
checks that need to be performed in addition to that one:

1. Insufficient authorization.
2. Insufficient authentication.
3. Insufficient encryption key size.
4. Insufficient encryption (check if persisted bond exists for
   unencrypted peer).
5. Invalid handle.
6. Write not permitted.

The host does not properly handle 3 and 4 yet (for any writes).

Now the host rejects a prepare write request if any of the above issues
are detected (except for 3 and 4).