BLE Host - sm,sc - ask for pk after rx pubkey
This bug occurs during secure connections pairing under the following
conditions:

* Initiator
* Pairing algorithm = passkey

The bug is that the host asked the application to input the passkey
before it was ready to accept it.  The host was asking for the passkey
immediately after sending its public key, but it isn't ready to accept
it until after the reponder sends its public key.  The fix is to wait
until receiving the responder's public key before asking for the
passkey.

This fix isn't ideal.  Ideally, the host would continue asking for the
passkey immediately, and would be able to accept it.  The easy
short-term fix is to hold off on asking the application.