BLE Host - Report updated MTU via GAP event cb.
Before this commit, MTU changes were only reported when we initiated the
exchange via a GATT procedure.  If the peer initiated the exchange, the
application never learned of an MTU update.

Now all MTU changes are reported through the GAP event callback.  The
callback passed to ble_gattc_exchange_mtu() is not very useful anymore.
Its only real purpose is to inform the user if the peer failed to
respond to the MTU request, which causes the connection to be terminated
anyway.

Relevant additions to the API:

 #define BLE_GAP_EVENT_MTU                   15

 struct ble_gap_event {
     /* [...] */

     /**
      * Represents a change in an L2CAP channel's MTU.
      *
      * Valid for the following event types:
      *     o BLE_GAP_EVENT_MTU
      */
     struct {
         /** The handle of the relevant connection. */
         uint16_t conn_handle;

         /**
          * Indicates the channel whose MTU has been updated; either
          * BLE_L2CAP_CID_ATT or the ID of a connection-oriented
          * channel.
          */
         uint16_t channel_id;

         /* The channel's new MTU. */
         uint16_t value;
     } mtu;
 };