BLE Host - Only log ATT cmd after tx succeeds.
Before, we were logging each ATT transmission as soon as a higher layer
requested the send.

Now:

* For client requests, don't log until after transmit succeeds.

* For server responses, don't log until we are sure we will attempt a
transmission.  Unfortunately, due to the way the server code is
structures, we don't have the necessary information for logging after
the transmit occurs.  This is still an improvement, as it prevents
spurious logging when there are invalid arguments or insufficient
resources.