MYNEWT-354
This fixes the compilation error when BLE_LL_CFG_FEAT_LE_ENCRYPTION
was set to 0. The issue here was that BLE_LL_CFG_FEAT_LE_PING cannot
be set to 1 if encryption is not defined (set to 0). Note that
LE_PING can be either defined or undefined (1 or 0) if encryption
is defined.