sim compiler - replace objsize with size
This appears to be a typo.  There is no tool called objsize.  size works
fine.

This only affects OS X.  It hasn't been a problem in the past when gcc-5
was used because the invocation of gobjdump was crashing with a message
like the following:

    2016/08/22 17:50:54 [DEBUG] o=gobjdump(68189,0x7fff71813000) malloc:
    *** error for object 0x1016002a0: pointer being freed was not
    allocated *** set a breakpoint in malloc_error_break to debug

When gobjdump crashed, newt ignored the error, but did not invoke the
size tool.

When clang is used instead of gcc, gobjdump doesn't crash, so the
incorrect size path is reported as a build failure.