MYNEWT-385: Controller not sending connection update complete event
The nimble controller was not sending the connection update
complete event if the connection was a slave, the host initiated
the connection parameter request and only an anchor point move was
requested. The spec is not perfectly clear but it seems this was a
bug and that the host should always be notified if it initiates
the connection update or connection parameter request.