console/full - Interpret '\n' as end-of-line.
The console history feature appears to have changed the EOL character
from '\n' to '\r'.  This caused problems for newtmgr, as the newtmgr
tool ends its commands with a '\n' character.

Now we interpret both '\r' and '\n' as EOL characters.  Brief testing
indicates that newtmgr and console history work with this change.