MYNEWT-397: Nimble controller assert if advertising started/stopped frequently
There was an assert in the advertising code in the nimble controller that
would assert if the advertising state machine was not enabled and an
advertising event finished. The code contained a race condition where
it was possible that the advertising state machine was stopped but
an advertising end event was processed by the Link Layer. The code will
now ignore advertising packets that are processed after advertising has
been stopped as these events are asynchronous to the controller.