Boot loader update
1. Increase max number of status entries

We weren't budgeting enough space for the swap status.  We were allowing
for 48 swap operations (32 * 4 / 3).  The nrf52dk, for example, requires
58 operations (232kB area / 4kB chunk = 58).

The result is that some status bytes could get copied to the opposing
slot.  On a subsequent swap, the status bytes would get copied back to
slot 0, corrupting the status.

2. Write status bytes in at increasing offsets

We were writing each subsequent status byte at an offset in flash
previous to the last.  Some flash hardware prohibits non-sequential
writes.

Still to do:

A. If the boot loader resets while it is in the middle of a revert (user
tests a new image, reboots, then reboots again), the behavior is
incorrect.  The image under test becomes confirmed - it should be the
original image that is confirmed.  The problem is that it is not
currently possible to determine the status of the previous swap
operation under these conditions.

Possible fix: when it appears there was no partial copy, search slot0's
status bytes for the swap status anyway.

B. If an image contains the "swap magic" (0x12344321) at an unfortunate
offset, the bootloader could get confused if it resets during a swap.
The boot loader.  I suggest we increase the size of the magic to 8 or 16
bytes.