bootutil - Only allocate sector arrays for loader.
The boot loader's array of sector descriptors was getting allocated
unconditionally, even though it was only used in the boot loader and in
split image code.  The problem is that the "-fdata-sections" gcc option
doesn't seem to have any effect.

The fix is to point to static arrays of sectors in a boot-loader-only
function.  These arrays get collected by the linker via
"-ffunction-sections".