MYNEWT-401: HAL timer init API modification
The changes are due to a modification of the hal timer API. We
decided that the configuration through the syscfg method was
not desirable and it was not consistent with other hals. Thus,
we modified hal_timer_init() to take a void *cfg and added a
hal_timer_config() to make it operate like the other HALs. There
are examples of configurations in most of the bsps for those who
want to see how to configure the platform specific *cfg that is
passed to hal_timer_init().