Move startup code back into BSP pgks (from MCU).
The MCU packages are probably the correct place for these files, but
newt has issues building split images with them there.  In particular,
newt doesn't recognize that the app (second slot) uses
Reset_Handler_split, so the resulting image is empty.  This is a
temporary fix to keep split images working.