MYNEWT-7 Add support for os_task_suspend
os_task_suspend() implements the ability to effectively stop a task.
A new routine os_sched_suspend() removes the task from the appropriate
scheduler queue and marks it as suspended. The task may be re-initialized
by calling os_task_init().

Suspending a task is disallowed if it's trying to suspend itself, it's not on the ready or sleep queue, or if it's waiting on a mutex, semaphore or eventq.