BLE Host - Eliminate spurious time expirations.
Prior to this change, the host timer would expire at least once per
second to clean up expired procedures.  This approach was used in lieu
of more diligent bookkeeping.  Functionally, this is OK, since these
timers don't need to be particularly precise.  However, this approach is
bad for power management, as it requires the processor to periodically
wake up from deep sleep when there is no work to do.