BLE Host - On rx indicate, allocate rsp first.
Don't process the indicate command until after we have successfully
allocated a response.  If response allocation fails, then reply with an
"insufficient resources" ATT error and throw away the indication.
Indicates are a bit different from other ATT commands because the
request is forwarded to the application as a GAP event, so we can't
reuse the request mbuf in the response.