MYNEWT-245 Allow debug of unittest via "newt run"
This change is required to enable the sibling change to work to function
properly.  With both changes, a unittest package can be debugged via:

    newt run <unittest-pkg>

This builds the test executable such that a failed test triggers a
failed assert and runs it in gdb.  When a test fails, gdb halts at the
point of failure.