nimble/test: Fix ble_att_svr_test_find_type_value test cases
Test cases for "Find By Type Value" are incorrect since they are based
on an assumption that attribute group is a sequence of consecutive
attributes with the same type and value.

This fixes the test cases to check for proper grouping as defined by
Bluetooth Specification 4.2, Vol 3, Part G, Section 2.5.3.

Since the actual ATT implementation is based on the same incorrect
assumtpion, the test cases will now fail until ATT is fixed as well.