No jira ticket
This is very, very initial BSP for the usbmkw41z. There is very limited HAL support and
the OS ticker will probably get changed as I did something simple for proof of concept.

Note that there is nothing in the devices directory in the SDK for this mcu. The reason
for this is that we currently do not have a good way of specifying which include paths
are used and there are identical files in each of the two devices currently in this
SDK. Thus, a number of the header files for this MCU have been moved to the mcu
specific directory (hw/mcu/nxp/mkw41z).