MYNEWT-524 Initial support for nrf52840pdk
This commit is the initial support for the nrf52840 pdk. I did some basic
ble testing with a test app (bletest) and it appears to be working. There
was no additional MCU added as the peripherals are close to the nrf52832.
No testing was done on any gpio outside the old 0 - 31 range.

Some changes were made that could affect the nrf52832 as well although
I doubt anything that changed with this commit would have any impact.