No jira ticket: nrf51 bsps should use #define for RTC
The nrf51 code should have used a #define for the os tick timer as
opposed to always using RTC0. I updated the code to use a #define
and also RTC1 as RTC0 may need to be used exclusively for the BLE
stack (not sure about that). Also, added a clock start function
to hal_system.c as I am not sure of ordering in regards to when
hal_bsp_init() will be called and when the clocks were started
by os_tick_init(). The new function checks to see if the desired
clock has already been started and if it has, do nothing.