split - Use _start_split symbol in app.
If the loader and app both use the same start symbol (_start), the
loader's main() gets used by the app, rather than the app's main.  This
causes the linker to strip pretty much all functionality from the app.

The solution is to use a different start symbol for the "application"
half of a split image.  Rather than _start, the app uses _start_split.
In addition, due to the way newt builds split images, _start_split has
to reside in a package that the loader doesn't use.  If it is in a
shared package, the entire package gets put in the loader, and
_start_split would erroneously reference the loader's main().