boot - Don't read full sector layout upfront.
Previously, the boot loader would read the properties of all image
sectors at startup.  This was a problem for the split image setup
because it required the allocation of a massive chunk of memory to hold
the sector info (at least 4kB).

Now, the boot loader retrieves sector information from the HAL on an
as-needed basis, eliminating the need for the upfront allocation.