MYNEWT-538 Add RTC timer to hal timers for nordic
There are two changes to the code. One was a bug fix where we need
to call the timer callback if we are within 3 ticks of a timer
expiration if the timer is an RTC timer. The other change was just
a modification of the code when checking for a timer that is too
far in the future.