bleprph - Move stack from bss to heap.
When an app is used as a loader, its bss contents are reserved and
off-limits to the application slot.  It is wasteful to put the
loader-objects big objects in bss, since they never get used anyway once
the loader jumps to the application-app.

The fix is to malloc the stack in the loader.