MYNEWT-568 log - Increase index to uint32 (was:16)
For the most part, this change is backwards compatible. The new behavior
is:

    /* If   lo_ts == -1: Only access last log entry;
     * Elif lo_ts == 0:  Don't filter by timestamp;
     * Else:             Only access entries whose ts >= lo_ts.
     */
    int64_t lo_ts;

    /* Only access entries whose index >= lo_index. */
    uint32_t lo_index;

So, the timestamp is still used as the primary key if a nonzero value is
specified. If 0 is specified, only the index is used to filter entries.

There is one compatibility-breaking change: the "logs show" command
retrieves all entries whose index is greater than or equal to the
specified index value. Prior to this change, the command only retrieved
entries whose index is greater. This change was made so that there is a
reliable means of getting an entry with an index of 0.