Revert "boot - Don't read full sector layout upfront."
This reverts commit e7fc231a687dabf70b53be004c18289b40fd876b

The original change was made to accommodate split image.  By not caching
sector information, the second app no longer had to malloc a large chunk
of memory.

It turns out there is an easier fix: the second app doesn't read sector
information in the first place.  This change will be made in a
subsequent commit.  The original design that caches sector information
is more straightforward for the boot loader, so the no-cache change is
being reverted.