MYNEWT-630: Connection supervision timeout with high throughput
The issue was found to be an error in the os scheduler code.
If the PendSV handler was interrupted at a certain point when
we were checking to see if a context switch could occur, the
code is os_sched would think no context switch was necessary.

This is now fixed and the high throughput test has been running
for quite some time now with no disconnect.