MYNEWT-669 NMP: Make max chunk size configurable.
Currently, the largest image or file chunk that a Mynewt device can
receive is hardcoded at 400 bytes. Furthermore, the largets CBOR
attribute that we can decode is 300 bytes long.This is not great because
this number is independent of the transport MTU, so the client has no
way of knowing the limit.

The fix is to create three new compile-time settings:

    * CBORATTR_MAX_SIZE
    * FS_UPLOAD_MAX_CHUNK_SIZE
    * IMGMGR_MAX_CHUNK_SIZE

and set them to 512 by default. This value is large enough to
accommodate full-size BLE packets.

This is still not a perfect solution because these values are not tied
to the MTU in any way. One way to truly solve this would be to have an
"NMP MTU" that the client can discover. That would be a pretty big
change.