MYNEWT-697: Controller can leak connection state machines
Fixed bug where it was possible that a controller could allocate
a connection state machine and never make it active or free it.
If this happens, it is possible that the controller would not
be able to create a new connection. This particular bug would
only occur if the device was a central and would occur during
the le create connection command.