nimble/controller: Fix for directed advertising report
Whenever scan filtering is 0x02 and initA is not resolved RPA,
we should send to host LE Direct Advertising Report event instead of
LE Advertising Report.
If InitA is not a RPA we should rather drop the adv packet.

With this patch we also stop to determine subevent type on own_address_type
as this is not related.