MYNEWT-668: Improve throughput of BLE connections
We will now send the number of completed packets event faster
than it was being sent before, assuming that more than 2 packets
were completed in a connection event. The old code would send
the event at the end of the connection event and this impacted
single connection throughput when trying to max out the
throughput.