sys/console: switch to blocking mode for tab completion
This patch solves an issue with tab completion and async console.
Tab completion is triggered in interrupt context and it outputs
text into the buffer. As the buffer fills up, the console tries to
wait until the uart pulls out the data from the buffer, but since
it all happens in the interrupt context, the uart cannot issue another
interrupt to pull the data thus freezing the console.

To avoid this we switch the console to blocking mode until the
completion is finished.