nimble/transport; MYNEWT-737: Make ble_hci_uart_tx_char more efficient
Modified the code in ble_hci_uart_tx_char to be faster. This
added a bit more code (should not have increased ram usage) and
should in general be faster than the old method and should also
free the mbufs slightly faster (each one is freed as it is
used).