nimble/ll: Add option to schedule connection with non-zero WindowOffset
Devices based on TI CC254x with older stack version have bug where
connection parameters are not updated properly on slave device if
transmitWindowOffset in CONNECT_IND was set to 0 (and apparently this
applies to 1 as well).

To workaround this, we allow to use larger transmitWindowOffset in
CONNECT_IND - by default this is set to 0, for old TI CC254x IOP should
be set to at least 2.

This also fixes MYNEWT-526.