nimble/controller: Cleanup PHY symbols names
To avoid confusion which phy-value goes where, let's use following
terms in LL code:
- phy = numeric PHY value (1M, 2M, Coded => 1, 2, 3)
- phy_mask = PHY bitmask (1M, 2M, Coded -> 0x01, 0x02, 0x04)
- phy_mode = actual PHY used in radio (1M, 2M, Coded S=2, Coded S=8)

Values for phy and phy_mask are the same as used in HCI and LL so can
be exchanged between layers.

Value for phy_mode is calculated from phy and phy_options using helper.

This patch adjusts various symbols names to the above, so in case e.g.
phy-variable is assigned to phy_mode-variable directly somewhere, this
potentially could indicate there is some issue in code.