MYNEWT_689: Make CONSOLE_UART a syscfg variable.
Removed CONSOLE_UART from bsp.h and made it a syscfg variable.
The variable is called CONSOLE_UART_DEV and is defined in the
sys/console directories. This can be overridden by the bsp, app,
or target. Some bsps used different settings for the console uart
and those bsps override the default uart (which is uart0).