hw: mcu: pic32mz2048efg100: Reduce number of UART TX interrupts
The UART TXIF flag is set when there is some space in the TX FIFO.
In the current implementation, only one byte is sent every time
the interrupt is handled.
This change modifies the behaviour of the UART TX interrupt
handler by sending as many bytes as possible until no more bytes
are available or the TX FIFO is full.

Signed-off-by: Francois Berder <fberder@outlook.fr>